+++ Eventodor/Screens/EventRegCard/Cells/DescriptionCell/DescriptionTableViewCell.swift
@@ -1,5 +1,5 @@
 //
-//  CardCellTableViewCell.swift
+//  DescriptionTableViewCell.swift
 //  Eventodor
 //
 //  Created by Petrykevich, Kanstantsin on 24.05.22.
@@ -7,16 +7,11 @@
 
 import Foundation
 import UIKit
-import Kingfisher
 
-class CardCellTableViewCell: UITableViewCell {
+class DescriptionTableViewCell: UITableViewCell {
     
     // MARK: - Outlets
-    @IBOutlet weak var cardImageView: UIImageView!
-    @IBOutlet weak var eventNameLabel: UILabel!
-    @IBOutlet weak var distanceLabel: UILabel!
-    @IBOutlet weak var ratingLabel: UILabel!
-    @IBOutlet weak var participantCountLabel: UILabel!
+    @IBOutlet weak var descriptionLabel: UILabel!
     @IBOutlet weak var timeTitleLabel: UILabel!
     @IBOutlet weak var timeLabel: UILabel!
     @IBOutlet weak var priceTitleLabel: UILabel!
@@ -34,11 +29,7 @@ class CardCellTableViewCell: UITableViewCell {
     // MARK: - Helper method's
     func configure(with event: Event?) {
         guard let event = event else { return }
-        cardImageView.kf.setImage(with: URL(string: event.photo?.url ?? ""))
-        eventNameLabel.text = event.name
-        distanceLabel.text = "\(event.distance ?? 0.0) km"
-        ratingLabel.text = event.rank ?? "" + " " + "*"
-        participantCountLabel.text = "\(event.register_persons_amount ?? 0)/\(event.persons_amount ?? 0)"
+        descriptionLabel.text = event.description
         timeTitleLabel.text = "time_title".localized()
         timeLabel.text = event.time ?? ""
         priceTitleLabel.text = "price_title".localized()
@@ -48,7 +39,7 @@ class CardCellTableViewCell: UITableViewCell {
 
 // MARK: - Private method's
 private
-extension CardCellTableViewCell {
+extension DescriptionTableViewCell {
     
     func setupUI() {
         setupLabels()
